feat: ✨ Replaced the default homepage with the invite interface
