+++ apps/wizarr-frontend/src/modules/home/events.ts
