+++ apps/wizarr-frontend/src/modules/home/pages/Emby/Signup.vue
