+++ apps/wizarr-frontend/src/modules/home/pages/Error.vue
