+++ apps/wizarr-frontend/src/modules/home/pages/Jellyfin/Signup.vue
