+++ apps/wizarr-frontend/src/modules/home/pages/Plex/Signup.vue
