+++ apps/wizarr-frontend/src/modules/home/pages/Stepper.vue
