+++ apps/wizarr-frontend/src/modules/home/router/index.ts
@@ -2,7 +2,23 @@ import type { RouteRecordRaw } from 'vue-router';
 
 const routes: Readonly<RouteRecordRaw[]> = [
     {
+        path: '/i/:invite',
+        name: 'invite',
+        component: () => import('../views/Home.vue'),
+    },
+    {
         path: '/',
+        name: 'home',
+        component: () => import('../views/Home.vue'),
+    },
+    // TODO: Remove this route after a few versions to allow users to get used to the new route
+    {
+        path: '/j/:invite',
+        name: 'invite',
+        component: () => import('../views/Home.vue'),
+    },
+    {
+        path: '/join',
         name: 'home',
         component: () => import('../views/Home.vue'),
     },
