+++ apps/wizarr-frontend/src/modules/home/types/JellyfinForm.ts
