--- apps/wizarr-frontend/src/modules/join/router/index.ts
@@ -1,16 +1,0 @@
-import type { RouteRecordRaw } from 'vue-router';
-
-const routes: Readonly<RouteRecordRaw[]> = [
-    {
-        path: '/j/:invite',
-        name: 'join-invite',
-        component: () => import('../views/Join.vue'),
-    },
-    {
-        path: '/join',
-        name: 'join',
-        component: () => import('../views/Join.vue'),
-    },
-];
-
-export default routes;
